refactor value module layout
